Rename launch info
Rename the launch info

Nits
