+++ src/omnisharp/OmniSharpLaunchInfo.ts
