+++ test/unitTests/GetOmniSharpLaunchInfo.test.ts
@@ -3,12 +3,12 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { GetOmniSharpLaunchInfo } from "../../src/omnisharp/GetOmniSharpLaunchInfo";
+import { getOmniSharpLaunchInfo } from "../../src/omnisharp/GetOmniSharpLaunchInfo";
 import { PlatformInformation } from "../../src/platform";
 import { expect } from "chai";
 import * as path from "path";
 
-suite(GetOmniSharpLaunchInfo.name, () => {
+suite(getOmniSharpLaunchInfo.name, () => {
     const basePath = "basePath";
 
     [
@@ -16,7 +16,7 @@ suite(GetOmniSharpLaunchInfo.name, () => {
         new PlatformInformation("win32", "x86_64"),
     ].forEach((platformInfo: PlatformInformation) => {
         test(`OmniSharp.exe is appended to the basepath and MonoLaunchPath is undefined for windows platform ${platformInfo.toString()}`, () => {
-            let launchInfo = GetOmniSharpLaunchInfo(platformInfo, basePath);
+            let launchInfo = getOmniSharpLaunchInfo(platformInfo, basePath);
             expect(launchInfo.MonoLaunchPath).to.be.undefined;
             expect(launchInfo.LaunchPath).to.equal(path.join(basePath, "OmniSharp.exe"));
         });
@@ -26,7 +26,7 @@ suite(GetOmniSharpLaunchInfo.name, () => {
         new PlatformInformation("linux", "x86_64"),
     ].forEach((platformInfo: PlatformInformation) => {
         test(`OmniSharp.exe is appended to the basepath and MonoLaunchPath is undefined for non windows platform`, () => {
-            let launchInfo = GetOmniSharpLaunchInfo(platformInfo, basePath);
+            let launchInfo = getOmniSharpLaunchInfo(platformInfo, basePath);
             expect(launchInfo.MonoLaunchPath).to.to.equal(path.join(basePath, "omnisharp" ,"OmniSharp.exe"));
             expect(launchInfo.LaunchPath).to.equal(path.join(basePath, "run"));
         });
